app/requirements.txt
flask
azure-ai-openai
azure-core